params: move five-piece support into extension tree
